+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/RoutingSource.kt
@@ -10,11 +10,14 @@ import io.reactivex.Observable
 import io.reactivex.ObservableSource
 import io.reactivex.Observer
 import io.reactivex.functions.BiFunction
-import java.util.ArrayDeque
 
 interface RoutingSource<C : Parcelable> :
     ObservableSource<RoutingHistory<C>>,
     SubtreeBackPressHandler {
+
+    operator fun plus(other: RoutingSource<C>): RoutingSource<C> = Combined(this, other)
+
+    fun remove(identifier: Routing.Identifier)
 
     data class Combined<C : Parcelable>(
         val first: RoutingSource<C>,
@@ -53,10 +56,6 @@ interface RoutingSource<C : Parcelable> :
         override fun handleBackPressFallback(): Boolean =
             first.handleBackPressFallback() || second.handleBackPressFallback()
     }
-
-    operator fun plus(other: RoutingSource<C>): RoutingSource<C> = Combined(this, other)
-
-    fun remove(identifier: Routing.Identifier)
 
     // TODO extract
     class Permanent<C : Parcelable>(permanents: Iterable<C>) : RoutingSource<C> {
@@ -102,40 +101,3 @@ interface RoutingSource<C : Parcelable> :
     }
 }
 
-// TODO extract
-// https://dev.to/alediaferia/a-concatenated-iterator-example-in-kotlin-1l23
-class ConcatIterator<T>(iterator: Iterator<T>) : Iterator<T> {
-    private val store = ArrayDeque<Iterator<T>>()
-
-    init {
-        if (iterator.hasNext())
-            store.add(iterator)
-    }
-
-    override fun hasNext(): Boolean = when {
-        store.isEmpty() -> false
-        else -> store.first.hasNext()
-    }
-
-    override fun next(): T {
-        val t = store.first.next()
-
-        if (!store.first.hasNext())
-            store.removeFirst()
-
-        return t
-    }
-
-    operator fun plus(iterator: Iterator<T>): ConcatIterator<T> {
-        if (iterator.hasNext())
-            store.add(iterator)
-        return this
-    }
-}
-
-operator fun <T> Iterator<T>.plus(iterator: Iterator<T>): ConcatIterator<T> =
-    when {
-        this is ConcatIterator<T> -> this.plus(iterator)
-        iterator is ConcatIterator<T> -> iterator.plus(this)
-        else -> ConcatIterator(this).plus(iterator)
-    }
